Random seed updates for experiement running scripts

+++ assignment1/run_experiment.py
@@ -39,7 +39,7 @@ if __name__ == '__main__':
 
     seed = args.seed
     if seed is None:
-        seed = np.random.randint(0, (2 ** 32) - 1)
+        seed = np.random.randint(0, (2 ** 32) - 1, dtype='uint64')
         print("Using seed {}".format(seed))
 
     print("Loading data")
